+++ Examples/HelloWorld/Package.swift
@@ -11,7 +11,7 @@ let package = Package(
         .library(name: "HelloWorld", targets: ["HelloWorld"]),
     ],
     dependencies: [
-        .package(url: "https://github.com/1amageek/swift-web.git", from: "0.10.0"),
+        .package(url: "https://github.com/1amageek/swift-web.git", from: "0.11.0"),
         .package(url: "https://github.com/1amageek/swift-html.git", from: "0.14.0"),
     ],
     targets: [